Initial commit
`Tokenizer` can _just barely_ parse a basic, well-formed move list.

Initially, I wanted to provide the movetext as a `Stream` rather than a
string, the idea being that it could be processed as it was being read
from a file without having to read the entire file into memory first. I
had difficulties with the stream being unreadable in
`Tokenizer.ParseMoves()`, so I switched to a string in order to get the
actual parsing logic down first.

Because of the `yield return` strategy, the debug console output
includes all of the expected halfmoves multiple times in various orders.
After running a test, generally the full, in-order list seems to exist
at the bottom of the output.